Moved create_img_from_text function to utils.py
