+++ src/class_cvae/verify_latent_effectiveness.py
@@ -15,24 +15,12 @@ from PIL import Image, ImageDraw, ImageFont
 import matplotlib.pyplot as plt
 
 from models import Encoder, Decoder, Classifier, ImageClassifier
-
+from utils import create_img_from_text
 
 def load_data():
     test_dset = MNIST(root="data", train=False, transform=ToTensor())
 
     return test_dset
-
-def create_img_from_text(width, height, text):
-    PAD = 2
-    img_size = img.shape[:2]
-    text_img = (np.ones((height, width, 3)) * 255).astype(np.uint8)
-    text_img = Image.fromarray(text_img)
-    text_img_dr = ImageDraw.Draw(text_img)
-    font = ImageFont.load_default()
-    text_img_dr.text((PAD, PAD), text, font=font, fill=(0, 0, 0))
-    text_img = np.array(text_img)[:, :, :1]
-
-    return text_img
 
 def save_imgs(reals, fakes, confs, org_confs, output):
     reals = reals.cpu().detach().numpy()
