+++ src/class_cvae/visual_counter_factual_split.py
@@ -15,7 +15,7 @@ from PIL import Image, ImageDraw, ImageFont
 import matplotlib.pyplot as plt
 
 from models import Encoder, Decoder, Classifier, ImageClassifier
-
+from utils import create_img_from_text
 
 """
 Goal: Create visual counterfactual
@@ -26,18 +26,6 @@ def load_data():
     test_dset = MNIST(root="data", train=False, transform=ToTensor())
 
     return test_dset
-
-def create_img_from_text(width, height, text):
-    PAD = 2
-    img_size = img.shape[:2]
-    text_img = (np.ones((height, width, 3)) * 255).astype(np.uint8)
-    text_img = Image.fromarray(text_img)
-    text_img_dr = ImageDraw.Draw(text_img)
-    font = ImageFont.load_default()
-    text_img_dr.text((PAD, PAD), text, font=font, fill=(0, 0, 0))
-    text_img = np.array(text_img)[:, :, :1]
-
-    return text_img
 
 def save_imgs(reals, fakes, cls_org, cls_fakes, confs, org_confs, output):
     reals = reals.cpu().detach().numpy()
